requirements.txt
Django==1.11.4
MySQL-python==1.2.5
Pillow==4.2.1
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
django-sql-explorer==1.1.1
idna==2.5
mysql==0.0.1
olefile==0.44
pytz==2017.2
requests==2.18.3
six==1.10.0
sqlparse==0.2.3
unicodecsv==0.14.1
urllib3==1.22
wsgiref==0.1.2
